Remove use respeaker as a variable
